restructure auth files
