+++ locally-app/app/(auth)/signup.tsx
@@ -1,9 +1,11 @@
 import { View, Text, Image, TouchableOpacity } from 'react-native'
 import React, { useState } from 'react'
-import { router } from 'expo-router'
+import { Link, router } from 'expo-router'
 import Ionicons from '@expo/vector-icons/Ionicons'
-import GoogleButton from '../components/GoogleButton'
-import FormInput from '../components/FormInput'
+import GoogleButton from '../../components/GoogleButton'
+import FormInput from '../../components/FormInput'
+
+import { images } from '@/constants'
 
 const PrimaryButton = ({ text, onPress }: { text: string; onPress: () => void }) => (
     <TouchableOpacity
@@ -21,33 +23,37 @@ const PrimaryButton = ({ text, onPress }: { text: string; onPress: () => void })
     </TouchableOpacity>
 )
 
-//Login logic goes here
-const LoginScreen = () => {
+const SignUpScreen = () => {
+    const [fullName, setFullName] = useState('');
     const [email, setEmail] = useState('');
     const [password, setPassword] = useState('');
 
-    const handleLogin = () => {
+    const handleSignUp = () => {
         //router.push('/(tabs)/metadata')
     }
 
     const handleGoogleLogin = () => {
     }
 
-
-
     return (
         <View className="flex-1 bg-white p-6">
             {/* Logo and Title */}
             <View className="items-center mb-8">
                 <Image
-                    source={require('../assets/images/splash.png')}
+                    source={ images.logo }
                     className="w-64 h-64"
                 />
-                <Text className="text-2xl font-bold text-center">Log In</Text>
+                <Text className="text-2xl font-bold text-center">Sign Up</Text>
             </View>
 
-            {/* Login Form */}
+            {/* Sign Up Form */}
             <View className="mb-6">
+                <FormInput
+                    icon="person-outline"
+                    placeholder="Full name"
+                    value={fullName}
+                    onChangeText={setFullName}
+                />
                 <FormInput
                     icon="mail-outline"
                     placeholder="abc@email.com"
@@ -62,17 +68,12 @@ const LoginScreen = () => {
                     onChangeText={setPassword}
                 />
 
-                <TouchableOpacity className="self-end mb-4">
-                    <Text className="text-orange-500">Forgot Password?</Text>
-                </TouchableOpacity>
-
                 <PrimaryButton
-                    text="LOG IN"
-                    onPress={handleLogin}
+                    text="REGISTER"
+                    onPress={handleSignUp}
                 />
             </View>
 
-            {/* Divider */}
             <View className="flex-row items-center mb-6">
                 <View className="flex-1 h-[1px] bg-gray-300" />
                 <Text className="mx-4 text-gray-500">OR</Text>
@@ -81,16 +82,16 @@ const LoginScreen = () => {
 
             <GoogleButton onPress={handleGoogleLogin} />
 
-
-            {/* Sign Up Link */}
+            {/* Login Link */}
             <View className="flex-row justify-center mt-6">
-                <Text className="text-gray-600">Don't have an account? </Text>
-                <TouchableOpacity onPress={() => router.push('/signup')}>
-                    <Text className="text-blue-500">Sign up</Text>
-                </TouchableOpacity>
+                <Text className="text-gray-600">Already have an account? </Text>
+                {/* <TouchableOpacity onPress={() => router.push('./login')}>
+                    <Text className="text-[#40BFFF]">Log In</Text>
+                </TouchableOpacity> */}
+                <Link href="./login">Log In</Link>
             </View>
         </View>
     )
 }
 
-export default LoginScreen
+export default SignUpScreen
